Create know-when-alt-text-should-be-left-blank.html
Know When Alt Text Should be Left Blank
In the last challenge, you learned that including an alt attribute when using img tags is mandatory. However, sometimes images are grouped with a caption already describing them, or are used for decoration only. In these cases, alt text may seem redundant or unnecessary.