Create wrap-radio-buttons-in-a-fieldset-element-for-better-accessibility.html
Wrap Radio Buttons in a fieldset Element for Better Accessibility
The next form topic covers the accessibility of radio buttons. Each choice is given a label with a for attribute tying to the id of the corresponding item as covered in the last challenge. Since radio buttons often come in a group where the user must choose one, there's a way to semantically show the choices are part of a set.